kmrl-app/backend/Rag-Engine/backend/requirements_fastapi.txt
# FastAPI RAG Engine Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0,<3.0.0
google-generativeai>=0.3.0
opensearch-py>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
scikit-learn>=1.0.0
python-multipart>=0.0.6
